Add conf modal

+++ src/app/administration/authorized-emails/authorized-emails.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { Subscription, catchError, tap, throwError } from 'rxjs';
+import { ConfirmModalComponent } from 'src/app/components/confirm-modal/confirm-modal.component';
 import { AuthorizedEmailsService } from 'src/app/services/authorized-emails.service';
 import { IconService } from 'src/app/services/icon.service';
 import { AccessRequest } from 'src/app/types/access-request';
@@ -15,7 +17,8 @@ export class AuthorizedEmailsComponent implements OnInit, OnDestroy {
 
   constructor(
     private authorizedEmailsService: AuthorizedEmailsService,
-    public icon: IconService
+    public icon: IconService,
+    private modalService: NgbModal
   ) {
     this.loadRequests();
     this.requestsSub = this.authorizedEmailsService.requestsChanged.subscribe(
@@ -42,17 +45,25 @@ export class AuthorizedEmailsComponent implements OnInit, OnDestroy {
   ngOnInit(): void {}
 
   protected onDeleteAuthorization(request: AccessRequest) {
-    this.authorizedEmailsService
-      .deleteRequest(request.id)
-      .pipe(
-        tap(() => {
-          this.loadRequests();
-          this.authorizedEmailsService.requestsChanged.next([request.id]);
-        }),
-        catchError((err) => {
-          return throwError(err);
-        })
-      )
-      .subscribe();
+    const modalRef = this.modalService.open(ConfirmModalComponent);
+
+    modalRef.componentInstance.message = `Are you sure you want to delete this authorization? ${request.name} will not be able to login or create an account with ${request.email} until you authorize it again.`;
+
+    modalRef.result.then((result) => {
+      if (result === 'Yes') {
+        this.authorizedEmailsService
+          .deleteRequest(request.id)
+          .pipe(
+            tap(() => {
+              this.loadRequests();
+              this.authorizedEmailsService.requestsChanged.next([request.id]);
+            }),
+            catchError((err) => {
+              return throwError(err);
+            })
+          )
+          .subscribe();
+      }
+    });
   }
 }
